requirements.txt
# Core
anthropic>=0.39.0
click>=8.1
jinja2>=3.1
pyyaml>=6.0
python-dotenv>=1.0
radon>=6.0

# Parsing
tree-sitter>=0.23.0
tree-sitter-javascript>=0.23.0
tree-sitter-typescript>=0.23.0

# Output
mkdocs>=1.5

# Dev
pytest>=7.0
pytest-cov>=4.0
ruff>=0.1
pre-commit>=3.0
